+++ docsrc/source/changelog.rst
@@ -3,6 +3,15 @@ Changelog
 
 Here you will find a list of changes for each package update.
 
+v0.27.0
+--------
+
+- Fixed datasets.documents.update_where so it runs
+- Added more tests around multivector search
+- Added Pandas-like Dataset Class for interacting with SDK (Alpha)
+- Added datasets.cluster.centroids.list_furthest_from_centers and datasets.cluster.centroids.list_closest_to_centers
+- Folder Refactor
+
 v0.26.6
 --------